+++ tests/mock/MockServer.h
@@ -46,7 +46,6 @@ typedef __int32 int32_t;
 
 #include <deflect/mock/api.h>
 #include <deflect/config.h>
-#include <deflect/NetworkProtocol.h>
 
 #include <QtNetwork/QTcpServer>
 
@@ -55,8 +54,7 @@ class MockServer : public QTcpServer
     Q_OBJECT
 
 public:
-    DEFLECT_API explicit MockServer( int32_t protocolVersion =
-                                     NETWORK_PROTOCOL_VERSION );
+    DEFLECT_API explicit MockServer( int32_t protocolVersion );
     DEFLECT_API virtual ~MockServer();
 
 protected: